requirements.txt
ultralytics
flask
opencv-python
pillow

